+++ templates/element/admin/sidebar.php
@@ -11,9 +11,9 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ *
+ * @var \MeCms\View\View\AdminView $this
  */
-
-/** @var \MeCms\View\View\AdminView $this */
 
 use Cake\Utility\Inflector;
 use MeCms\Core\Plugin;