+++ templates/element/admin/userbar.php
@@ -11,13 +11,11 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ *
+ * @var \MeCms\View\View\AdminView $this
  */
 
 use MeCms\Core\Plugin;
-
-/**
- * @property \MeTools\View\Helper\BootstrapDropdownHelper $Dropdown
- */
 
 $this->extend('MeCms.common/userbar');
 ?>