+++ templates/element/common/userbar.php
@@ -11,9 +11,6 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- *
- * This element is extended by `templates/element/userbar.php` for the frontend
- *  userbar and by `templates/element/admin/userbar.php` for the backend userbar.
  */
 ?>