+++ templates/element/topbar.php
@@ -11,15 +11,16 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ *
+ * @var \MeCms\View\View\AppView $this
  */
 
 use Cake\Core\App;
 use MeCms\Core\Plugin;
 
 /**
- * The topbar element will use the `TopbarHelper` from APP to build links, if
- *  that helper exists. Otherwise it will use the helper provided by MeCms, with
- *  the helper of any other plugin.
+ * The topbar element will use the `TopbarHelper` from APP to build links, if that helper exists. Otherwise, it will use
+ *  the helper provided by MeCms, with the helper of any other plugin.
  */
 ?>
 
@@ -36,14 +37,18 @@ use MeCms\Core\Plugin;
     <div class="container collapse navbar-collapse" id="topbarNav">
         <?php
         $app = (bool)App::className('TopbarHelper', 'View/Helper');
-        $links = $this->loadHelper($app ? 'Topbar' : 'MeCms.Topbar')->build();
+        /** @var \MeCms\View\Helper\TopbarHelper $TopbarHelper */
+        $TopbarHelper = $this->loadHelper($app ? 'Topbar' : 'MeCms.Topbar');
+        $links = $TopbarHelper->build();
         $this->helpers()->unload('Topbar');
 
         if (!$app) {
             //Builds links with any other plugin helper
             foreach (Plugin::all(['core' => false, 'exclude' => ['MeCms']]) as $plugin) {
                 if (App::className($plugin . '.TopbarHelper', 'View/Helper')) {
-                    $links = [...$links, ...$this->loadHelper($plugin . '.Topbar')->build()];
+                    /** @var \MeCms\View\Helper\TopbarHelper $TopbarHelper */
+                    $TopbarHelper = $this->loadHelper($plugin . '.Topbar');
+                    $links = [...$links, ...$TopbarHelper->build()];
                     $this->helpers()->unload('Topbar');
                 }
             }